django-backend/requirements.txt
asgiref==3.11.0
black==25.12.0
click==8.3.1
flake8==7.3.0
isort==7.0.0
mccabe==0.7.0
mypy_extensions==1.1.0
packaging==25.0
pathspec==1.0.3
platformdirs==4.5.1
pycodestyle==2.14.0
pyflakes==3.4.0
pytokens==0.3.0
sqlparse==0.5.5
Django>=4.2,<5.0
djangorestframework
gunicorn
Django>=4.2
WeasyPrint>=59.0


